+++ generators/server/files-sql.mjs
@@ -16,7 +16,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { mergeSections, addSectionsCondition, moveToJavaPackageSrcDir, moveToJavaPackageTestDir } from './support/utils.mjs';
+import { mergeSections, addSectionsCondition } from '../base/support/index.mjs';
+import { moveToJavaPackageSrcDir, moveToJavaPackageTestDir } from './support/index.mjs';
 import { SERVER_MAIN_SRC_DIR, SERVER_MAIN_RES_DIR, SERVER_TEST_SRC_DIR, SERVER_TEST_RES_DIR } from '../generator-constants.mjs';
 
 export const sqlFiles = {
